Fix some example C++ Warnings

+++ examples/core/core_drop_files.cpp
@@ -48,7 +48,7 @@ int main() {
                 raylib::DrawText("Dropped files:", 100, 40, 20, DARKGRAY);
 
                 // Iterate through all the dropped files.
-                for (int i = 0; i < droppedFiles.size(); i++) {
+                for (int i = 0; i < (int)droppedFiles.size(); i++) {
                     if (i % 2 == 0)
                         DrawRectangle(0, 85 + 40*i, screenWidth, 40, Fade(LIGHTGRAY, 0.5f));
                     else
